requirements.txt
protobuf==4.25.3
tensorflow==2.12.0
mediapipe==0.10.21
opencv-python
numpy
pillow
fastapi
uvicorn[standard]
python-multipart
streamlit